@@ -1,14 +1,14 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <TargetFramework>net5.0</TargetFramework>
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="Octokit" Version="0.47.0" />
-      <PackageReference Include="Serilog" Version="2.9.1-dev-01177" />
+      <PackageReference Include="Octokit" Version="0.50.0" />
+      <PackageReference Include="Serilog" Version="2.10.0" />
       <PackageReference Include="Serilog.Sinks.Console" Version="4.0.0-dev-00839" />
       <PackageReference Include="Serilog.Sinks.File" Version="5.0.0-dev-00876" />
       <PackageReference Include="System.CommandLine" Version="2.0.0-beta1.20253.1" />
